Organized the code into diff files
