+++ idea1/Section 3/AI_Images.html
@@ -50,6 +50,7 @@
         <div class="sentence-seg" id="textarea-container"></div>
         <button type="submit" class="SubmitButton">Submit</button>
     </form>
-    <script src="Image_scripts.js"></script>
+    <!-- Link the main JavaScript file -->
+    <script type="module" src="main.js"></script>
 </body>
 </html>
