+++ idea1/Section 3/main.js
@@ -1,0 +1,17 @@
+import { createElementsFromSentence, resetSentence } from './elements.js';
+import { addEventListeners } from './events.js';
+import { initializeModal } from './modal.js';
+
+const sentence = "This is sentence";
+
+document.addEventListener('DOMContentLoaded', () => {
+    createElementsFromSentence(sentence);
+    addEventListeners();
+    initializeModal();
+
+    // Initially disable the submit button
+    const submitButton = document.querySelector('.SubmitButton');
+    submitButton.disabled = true;
+});
+
+export { sentence };
