+++ idea1/Section 3/modal.js
@@ -1,0 +1,26 @@
+export function initializeModal() {
+    const modal = document.getElementById("Help-Page");
+
+    // Get the button that opens the modal
+    const btn = document.getElementById("HelpBtn");
+
+    // Get the <span> element that closes the modal
+    const span = document.getElementsByClassName("closeBtn")[0];
+
+    // When the user clicks the button, open the modal 
+    btn.onclick = function() {
+        modal.style.display = "block";
+    }
+
+    // When the user clicks on <span> (x), close the modal
+    span.onclick = function() {
+        modal.style.display = "none";
+    }
+
+    // When the user clicks anywhere outside of the modal, close it
+    window.onclick = function(event) {
+        if (event.target == modal) {
+            modal.style.display = "none";
+        }
+    }
+}
